+++ components/portfolio/inventory/collection/InventoryCollectionSettings.tsx
@@ -1,7 +1,36 @@
+'use client';
 import MagnifyingGlass from '@/public/icon/MagnifyingGlass';
 import styles from './InventoryCollectionSettings.module.css';
 import Gear from '@/public/icon/Gear';
+import { useAtom } from 'jotai';
+import { ChangeEvent, useState } from 'react';
+import { inventoryCollectionAtom } from '@/store/requestParam';
+import { priceTypeAtom } from '@/store/currency';
+import ReactModal from 'react-modal';
+import SpamModal from './SpamModal';
 const InventoryCollectionSettings = () => {
+  const [inventoryCollection, setInventoryCollection] = useAtom(
+    inventoryCollectionAtom
+  );
+  const [showModal, setShowModal] = useState(false);
+  const [priceType, setPriceType] = useAtom(priceTypeAtom);
+  const [searchText, setSearchText] = useState<string>('');
+  const handleTogglePriceType = () => {
+    setPriceType(priceType === 'costBasis' ? 'acquisitionPrice' : 'costBasis');
+  };
+  const handleInputText = (e: ChangeEvent<HTMLInputElement>) => {
+    const { value } = e.target;
+    setSearchText(value);
+    (value.length >= 3 || value.length == 0) &&
+      setInventoryCollection({
+        ...inventoryCollection,
+        page: 0,
+        searchCollection: e.target.value,
+      });
+  };
+  const handleModalOpen = () => {
+    setShowModal(true);
+  };
   return (
     <section className={`${styles.container}`}>
       <div className={`${styles.inputContainer}  dark:border-border-main-dark`}>
@@ -14,6 +43,8 @@ const InventoryCollectionSettings = () => {
           type='text'
           placeholder={'Search collection'}
           className={`${styles.textInput} font-caption-regular placeholder:dark:text-text-subtlest-dark`}
+          onChange={handleInputText}
+          value={searchText}
         />
       </div>
       <div className='flex items-center'>
@@ -21,17 +52,42 @@ const InventoryCollectionSettings = () => {
           <p className={`${styles.pSetting} dark:text-text-subtle-dark`}>
             Include Gas fee
           </p>
-          <button className={styles.toggleButton}>
+          <button
+            className={styles.toggleButton}
+            onClick={() => handleTogglePriceType()}
+          >
             <div className={styles.toggleIcon} />
           </button>
         </div>
         <button
           className={`font-button03-medium ${styles.settingButton} dark:border-border-main-dark`}
+          onClick={() => handleModalOpen()}
         >
           <Gear className={`${styles.gearIcon} dark:fill-icon-subtle-dark`} />
           <p className='dark:text-text-subtle-dark'>Spam Settings</p>
         </button>
       </div>
+      <ReactModal
+        isOpen={showModal}
+        contentLabel='Minimal Modal Example'
+        className='w-fit absolute top-[20%] left-[30%]'
+        onRequestClose={() => {
+          setShowModal(false);
+        }}
+        ariaHideApp={false}
+        shouldCloseOnOverlayClick={true}
+        overlayClassName={'overlayBackground'}
+      >
+        <div className='relative w-fit h-fit'>
+          <button
+            className='absolute top-10 right-10'
+            onClick={() => setShowModal(false)}
+          >
+            close
+          </button>
+          <SpamModal />
+        </div>
+      </ReactModal>
     </section>
   );
 };
