+++ components/portfolio/inventory/collection/InventoryCollectionTable.module.css
@@ -13,12 +13,9 @@
   @apply py-10 border-b-1 border-border-disabled;
   @apply text-text-main;
 }
-.tableCell {
-  @apply px-20 py-16;
+.spamButton {
+  @apply p-4 border-1 rounded-none bg-transparent border-border-bold;
 }
-.tableCell2 {
-  @apply px-20 py-16 col-span-2;
-}
-.tableCell3 {
-  @apply px-20 py-16 col-span-3;
+.spamButton path {
+  @apply fill-icon-subtle;
 }
