+++ public/icon/DotsThree.tsx
@@ -1,0 +1,19 @@
+const DotsThree = (props: any) => {
+  return (
+    <svg
+      width='16'
+      height='16'
+      viewBox='0 0 16 16'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+      {...props}
+    >
+      <path
+        d='M9 8C9 8.19778 8.94135 8.39112 8.83147 8.55557C8.72159 8.72002 8.56541 8.84819 8.38268 8.92388C8.19996 8.99957 7.99889 9.01937 7.80491 8.98079C7.61093 8.9422 7.43275 8.84696 7.29289 8.70711C7.15304 8.56725 7.0578 8.38907 7.01922 8.19509C6.98063 8.00111 7.00043 7.80004 7.07612 7.61732C7.15181 7.43459 7.27998 7.27841 7.44443 7.16853C7.60888 7.05865 7.80222 7 8 7C8.26522 7 8.51957 7.10536 8.70711 7.29289C8.89464 7.48043 9 7.73478 9 8ZM3.75 7C3.55222 7 3.35888 7.05865 3.19443 7.16853C3.02998 7.27841 2.90181 7.43459 2.82612 7.61732C2.75043 7.80004 2.73063 8.00111 2.76922 8.19509C2.8078 8.38907 2.90304 8.56725 3.04289 8.70711C3.18275 8.84696 3.36093 8.9422 3.55491 8.98079C3.74889 9.01937 3.94996 8.99957 4.13268 8.92388C4.31541 8.84819 4.47159 8.72002 4.58147 8.55557C4.69135 8.39112 4.75 8.19778 4.75 8C4.75 7.73478 4.64464 7.48043 4.45711 7.29289C4.26957 7.10536 4.01522 7 3.75 7ZM12.25 7C12.0522 7 11.8589 7.05865 11.6944 7.16853C11.53 7.27841 11.4018 7.43459 11.3261 7.61732C11.2504 7.80004 11.2306 8.00111 11.2692 8.19509C11.3078 8.38907 11.403 8.56725 11.5429 8.70711C11.6827 8.84696 11.8609 8.9422 12.0549 8.98079C12.2489 9.01937 12.45 8.99957 12.6327 8.92388C12.8154 8.84819 12.9716 8.72002 13.0815 8.55557C13.1914 8.39112 13.25 8.19778 13.25 8C13.25 7.73478 13.1446 7.48043 12.9571 7.29289C12.7696 7.10536 12.5152 7 12.25 7Z'
+        fill='#9CA3AF'
+        {...props}
+      />
+    </svg>
+  );
+};
+export default DotsThree;
